requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
torch>=1.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
hyperliquid-python-sdk>=0.0.1
python-dotenv>=1.0.0
rich>=13.0.0
questionary>=2.0.0
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0